1. 将 operator new 参数 从 unsighed int 改为 unsigned long long 2. 增加排序类

+++ Object.cpp
@@ -8,7 +8,7 @@ namespace DTLib
 {
 
 // 异常规格说明, 当前函数不会抛出任何异常
-void* Object::operator new(unsigned int size) throw()
+void* Object::operator new(unsigned long long size) throw()
 {
     return malloc(size);
 }
@@ -18,7 +18,7 @@ void Object::operator delete(void* p)
     free(p);
 }
 
-void* Object::operator new[](unsigned int size) throw()
+void* Object::operator new[](unsigned long long size) throw()
 {
 
     return malloc(size);
